python: move sockets to their own modules
Discovered that the `#[pyclass]` proc-macro takes a `module` argument
that allows us to remove (most of) the annoying module structure I had
thought was required by pyo3 -- types now appear as part of the
`tailscale` module without having to be declared there. They still need
to be explicitly reexported from it, though.
